### 2018-04-13 822eead
[Docs] Minor fix in paper
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
