+++ docs/paper/paper.bib
@@ -93,7 +93,7 @@
 }
 
 @article{sinclair1990modified,
-	author = {C.D.Sinclair and B.D.Spurr and M.I.Ahmad},
+	author = {C.D. Sinclair and B.D. Spurr and M.I. Ahmad},
 	title = {Modified anderson darling test},
 	journal = {Communications in Statistics - Theory and Methods},
 	volume = {19},
