### 2018-04-22 #675 @danielskatz
@federeghe - please suggest potential reviewers. [This list of people](https://bit.ly/joss-reviewers) who have already agreed to review for JOSS may be candidates.
