### 2018-04-22 #675 @federeghe
Thank you, maybe @fritzo or @jordigh? According to the list, they are experts in C++ and statistics, not on real-time systems but it may be sufficient for the review.
