### 2018-05-02 #711 @whedon
[REVIEW]: chronovise: Measurement-Based Probabilistic Timing Analysis framework
**Submitting author:** @federeghe (<a href="http://orcid.org/0000-0002-1888-9579">FEDERICO REGHENZANI</a>)
**Repository:** <a href="https://github.com/federeghe/chronovise" target ="_blank">https://github.com/federeghe/chronovise</a>
**Version:** v1.0
