### 2018-05-07 #711 @lkosmid
@danielskatz - yes, indeed,  everything is ok. I think I will be able to start reviewing at the end of this week.
