### 2018-05-14 #711 @lkosmid
@danielskatz - review it's under way, prepared a clean environment to test the software
