### 2018-05-22 2997cb4
[Documentation] Fix README license information
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
