### 2018-05-22 f6018e7
[Documentation] Add status badge to README table
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
