### 2018-05-22 0df8c1d
[Documentation] Improve README installation and building instructions
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
