### 2018-05-23 4f6b5f3
[Misc] Cosmetics and minor fixes
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
