+++ include/statistical/test_ks.hpp
@@ -56,7 +56,7 @@ public:
     /**
      * @copydoc StatisticalTest::run()
      */
-    virtual void run(const MeasuresPool<T_INPUT, T_TIME> &measures);
+    virtual void run(const MeasuresPool<T_INPUT, T_TIME> &measures) override;
 
     /**
      * @copydoc StatisticalTest::get_power()
