+++ include/statistical/test_ljung_box.hpp
@@ -57,7 +57,7 @@ public:
     /**
      * @copydoc StatisticalTest::run()
      */
-    virtual void run(const MeasuresPool<T_INPUT, T_TIME> &measures);
+    virtual void run(const MeasuresPool<T_INPUT, T_TIME> &measures) override;
 
     /**
      * @copydoc StatisticalTest::get_power()
