### 2018-05-24 7714365
[Misc] Fix README compiler versions
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
