### 2018-05-27 #711 @danielskatz
👋 @lkosmid - it looks like the issues you brought up have been addressed

Please confirm, and please continue your review
