### 2018-05-28 #711 @lkosmid
@danielskatz - Indeed, thanks @federeghe
I continue with the review
