### 2018-06-01 #711 @lkosmid
In general the short paper seems adequate. It motivates the need for Measurement Based Probabilistic Timing Analysis and in particular the need for an open source tool to compute pWCETs. However, I think that some important references are missing as well as the implementation of some of hte a state-of-the-art MBPTA features.

You can find more details comments here: https://github.com/federeghe/chronovise/issues/11
