### 2018-06-01 4d8e521
[Statistical] Add skeleton for CV estimator class
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
