### 2018-06-05 756d7fe
[Statistical] Minor fixes in Estimator_CV
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
