### 2018-06-05 cff85a3
[Samples] Add sample for CV estimation
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
