+++ samples/CMakeLists.txt
@@ -15,3 +15,11 @@ target_link_libraries(simple_hello_world PUBLIC chronovise)
 if(COVERAGE)
         target_link_libraries(simple_hello_world PUBLIC --coverage)
 endif(COVERAGE)
+
+set(SHW_CV_SOURCES ${SHW_CV_SOURCES} simple_hello_world_cv.cpp simple_main_cv.cpp)
+add_executable(simple_hello_world_cv ${SHW_CV_SOURCES})
+target_link_libraries(simple_hello_world_cv PUBLIC chronovise)
+
+if(COVERAGE)
+        target_link_libraries(simple_hello_world_cv PUBLIC --coverage)
+endif(COVERAGE)
