### 2018-06-05 9f7ef59
[EVTApproach_CV] Minor fixes in distribution location
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
