### 2018-06-06 5710619
[EVTApproach] Minor fixes
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
