### 2018-06-06 641ea78
[AEC/HMI] Add print of CV value
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
