### 2018-06-06 c22d7ba
[Misc] Remove unneeded <iostream>
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
