### 2018-06-06 b751801
[Paper] Improvements after review
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
