+++ docs/paper/paper.md
@@ -33,8 +33,8 @@ typical real-time constrained applications. Time constraints in CPS are often
 mandatory requirements, i.e. the timing requirements must be satisfied in any
 condition because of the mission-critical system purpose. The satisfaction of
 these constraints is traditionally demonstrated using well-known *static
-analyses* providing the **Worst-Case Execution Time (WCET)**. However, the
-increasing complexity of computer architectures -- such as multi-core,
+analyses* providing the **Worst-Case Execution Time (WCET)** [@wilhelm2008the].
+However, the increasing complexity of computer architectures -- such as multi-core,
 multi-level caches, complex pipelines, etc. [@berg2004requirements] --  makes
 these analyses computationally unaffordable or too pessimistic. The problem
 grows when dealing with Commercial-Off-The-Shelf (COTS) hardware
@@ -42,34 +42,46 @@ grows when dealing with Commercial-Off-The-Shelf (COTS) hardware
 
 *Probabilistic hard real-time* has been proposed as a possible solution to
 address this complexity increase [@bernat2002wcet]. In particular, the
-**Measurement-Based Probabilistic Time Analysis (MBPTA)** is a probabilistic
-real-time branch focused on the estimation of the WCET, directly from the
-observed execution times of real-time tasks.  The task execution time is
-measured across its input domain and the WCET is provided as a so-called
-probabilistic-WCET (pWCET), i.e. a WCET with a probability of observing higher
-execution times.  The statistical **Extreme Value Theory (EVT)**, typically
-used in natural disaster risk evaluation, is the basis for pWCET estimation.
-However, to obtain a safe pWCET estimation, the execution time traces must
-fulfill the requirements of EVT that can be checked running appropriate testing
-procedures [@santinelli2017revising].
+**Measurement-Based Probabilistic Time Analysis (MBPTA)**
+[@grosjean2012measurement] is a probabilistic real-time branch focused on the
+estimation of the WCET, directly from the observed execution times of real-time
+tasks.  The task execution time is measured across its input domain and the WCET
+is provided as a so-called probabilistic-WCET (pWCET), i.e. a WCET with a
+probability of observing higher execution times. The statistical
+**Extreme Value Theory (EVT)** [@castillo2005extreme] [@de2007extreme],
+typically used in natural disaster risk evaluation,
+is the basis for pWCET estimation. However, to obtain a safe pWCET estimation,
+the execution time traces must fulfill the requirements of EVT that can be
+checked running appropriate testing procedures [@santinelli2017revising]. In
+particular, MBPTA requires the time measurements to be [@kosmidis2017enabling]:
+(1) independent and identically distributed, (2) representative of all
+worst-case latency. The first requirement can be relaxed under some
+circumstances [@santinelli2017revising], while the latter is relative to the
+input representativity and to the hardware properties. These two requirements
+is necessary to obtain a safe, i.e. non-underestimated, pWCET.
 
 The *chronovise* framework is an open-source software aiming at standardizing
 the flow of MBPTA process, integrating both estimation and testing phases. The
-few previous works in literature are [@5591317] and [@Lesage2015], for which
-source code is not available.  Moreover, both works include a limited set of
-features, other than poor maturity level due to the missing integration of the
-most recent scientific contributions.  Our work aim at filling the absence of
+few previous software in literature are cited in [@5591317] and [@Lesage2015],
+for which source code is not available. Moreover, both works include a limited
+set of features, other than poor maturity level due to the missing integration
+of the most recent scientific contributions. Another software is available as
+open-source [@abella_jaume_2017_1065776], but specialized for a variant of
[…]
