### 2018-06-06 01d8cd3
[Paper] Fix bibtex entries
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
