+++ docs/paper/paper.bib
@@ -1,116 +1,230 @@
 @inproceedings{berg2004requirements,
-	author =	{Christoph Berg and Jakob Engblom and Reinhard Wilhelm},
-	title =	{Requirements for and Design of a Processor with Predictable Timing},
-	booktitle =	{Perspectives Workshop: Design of Systems with Predictable Behaviour},
-	year =	{2004},
-	editor =	{Lothar Thiele and Reinhard Wilhelm},
-	number =	{03471},
-	series =	{Dagstuhl Seminar Proceedings},
-	ISSN =	{1862-4405},
-	publisher =	{Internationales Begegnungs- und Forschungszentrum f{\"u}r Informatik (IBFI), Schloss Dagstuhl, Germany},
-	address =	{Dagstuhl, Germany},
-	URL =		{http://drops.dagstuhl.de/opus/volltexte/2004/5}
+    author    = {Christoph Berg and Jakob Engblom and Reinhard Wilhelm},
+    title     = {Requirements for and Design of a Processor with Predictable Timing},
+    booktitle = {Perspectives Workshop: Design of Systems with Predictable Behaviour},
+    year      = {2004},
+    editor    = {Lothar Thiele and Reinhard Wilhelm},
+    number    = {03471},
+    series    = {Dagstuhl Seminar Proceedings},
+    ISSN      = {1862-4405},
+    publisher = {Internationales Begegnungs- und Forschungszentrum f{\"u}r Informatik (IBFI), Schloss Dagstuhl, Germany},
+    address   = {Dagstuhl, Germany},
+    URL       = {http://drops.dagstuhl.de/opus/volltexte/2004/5}
 }
 
 @article{cullmann2010predictability,
-	title={Predictability considerations in the design of multi-core embedded systems},
-	author={Cullmann, Christoph and Ferdinand, Christian and Gebhard, Gernot and Grund, Daniel and Maiza, Claire and Reineke, Jan and Triquet, Beno{\^\i}t and Wilhelm, Reinhard},
-	journal={Proceedings of Embedded Real Time Software and Systems},
-	pages={36--42},
-	year={2010}
+    title  = {Predictability considerations in the design of multi-core embedded systems},
+    author = {Cullmann, Christoph and Ferdinand, Christian and Gebhard, Gernot and Grund, Daniel and Maiza, Claire and Reineke, Jan and Triquet, Beno{\^\i}t and Wilhelm, Reinhard},
+    journal= {Proceedings of Embedded Real Time Software and Systems},
+    pages  = {36--42},
+    year   = {2010}
 }
 
 @inproceedings{reghenzani2017mixed,
-	author={F. Reghenzani and G. Massari and W. Fornaciari}, 
-	booktitle={2017 Euromicro Conference on Digital System Design (DSD)}, 
-	title={Mixed Time-Criticality Process Interferences Characterization on a Multicore Linux System}, 
-	year={2017},
-	pages={427-434},
-	doi={10.1109/DSD.2017.18},
-	month={Aug}
+    author={Federico Reghenzani and Giuseppe Massari and William Fornaciari},
+    booktitle={2017 Euromicro Conference on Digital System Design (DSD)},
+    title={Mixed Time-Criticality Process Interferences Characterization on a Multicore Linux System},
+    year={2017},
+    pages={427-434},
+    doi={10.1109/DSD.2017.18},
+    month={Aug}
 }
 
 @inproceedings{dasari2013identifying,
-	author={D. Dasari and B. Akesson and V. Nélis and M. A. Awan and S. M. Petters}, 
-	booktitle={2013 8th IEEE International Symposium on Industrial Embedded Systems (SIES)}, 
-	title={Identifying the sources of unpredictability in COTS-based multicore systems}, 
-	year={2013}, 
-	pages={39-48}, 
[…]
