### 2018-06-06 8cc8d01
[Paper] Massari review
Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>
