+++ docs/paper/paper.md
@@ -25,19 +25,20 @@ bibliography: paper.bib
 
 # Summary
 The rapid advance of computer architectures towards more powerful, but also
-more sophisticated and complex platforms, has the side effect of making the
+more complex platforms, has the side effect of making the
 timing analysis of applications a challenging task [@cullmann2010predictability].
 The increasing demand of computational power
 in cyber-physical systems (CPS) is getting hard to fulfill, if we consider
 typical real-time constrained applications. Time constraints in CPS are often
-mandatory requirements, i.e. the timing requirements must be satisfied in any
+mandatory requirements, i.e. they must be satisfied in any
 condition because of the mission-critical system purpose. The satisfaction of
-these constraints is traditionally demonstrated using well-known *static
-analyses* providing the **Worst-Case Execution Time (WCET)** [@wilhelm2008the].
-However, the increasing complexity of computer architectures -- such as multi-core,
+these constraints is traditionally demonstrated using well-established *static
+analyses*, providing the **Worst-Case Execution Time (WCET)** [@wilhelm2008the].
+However, the increasing complexity of computing architectures -- such as multi-core,
 multi-level caches, complex pipelines, etc. [@berg2004requirements] --  makes
-these analyses computationally unaffordable or too pessimistic. The problem
-grows when dealing with Commercial-Off-The-Shelf (COTS) hardware
+these analyses computationally unaffordable or carrying out too pessimistic
+approximations.
+The problem grows when dealing with Commercial-Off-The-Shelf (COTS) hardware
 [@dasari2013identifying] and complex operating systems [@reghenzani2017mixed].
 
 *Probabilistic hard real-time* has been proposed as a possible solution to
@@ -45,17 +46,18 @@ address this complexity increase [@bernat2002wcet]. In particular, the
 **Measurement-Based Probabilistic Time Analysis (MBPTA)**
 [@grosjean2012measurement] is a probabilistic real-time branch focused on the
 estimation of the WCET, directly from the observed execution times of real-time
-tasks.  The task execution time is measured across its input domain and the WCET
-is provided as a so-called probabilistic-WCET (pWCET), i.e. a WCET with a
-probability of observing higher execution times. The statistical
-**Extreme Value Theory (EVT)** [@castillo2005extreme] [@de2007extreme],
-typically used in natural disaster risk evaluation,
-is the basis for pWCET estimation. However, to obtain a safe pWCET estimation,
-the execution time traces must fulfill the requirements of EVT that can be
-checked running appropriate testing procedures [@santinelli2017revising]. In
+tasks.  The time samples are collected across the application input domain and the WCET
+is provided in probablistic terms, the *probabilistic-WCET (pWCET)*, i.e. a WCET with a
+probability of observing higher execution times. The statistical theory at
+the basis of the WCET estimation is the **Extreme Value Theory (EVT)**
+[@castillo2005extreme] [@de2007extreme], typically used in natural disaster
+risk evaluation,
+However, to obtain a safe pWCET estimation,
+the execution time traces must fulfill the requirements of EVT. This can be
+checked by running suitable testing procedures [@santinelli2017revising]. In
 particular, MBPTA requires the time measurements to be [@kosmidis2017enabling]:
 (1) independent and identically distributed, (2) representative of all
-worst-case latency. The first requirement can be relaxed under some
+worst-case latency.  The first requirement can be relaxed under some
 circumstances [@santinelli2017revising], while the latter is relative to the
 input representativity and to the hardware properties. These two requirements
 is necessary to obtain a safe, i.e. non-underestimated, pWCET.
@@ -68,11 +70,10 @@ set of features, other than poor maturity level due to the missing integration
[…]
