### 2018-06-06 #711 @federeghe
@lkosmid We updated the article and the code.

We replied inline to the issue federeghe/chronovise#11

Thank you again for the detailed review.
