### 2018-06-08 federeghe/chronovise#11 @lkosmid
Thank you for the paper changes.

Some minor additional changes in order to close this issue:

- Probabilistic hard real-time has been proposed -> Probabilistic approaches for hard real-time systems have been proposed
- MBPTA is a probabilistic real-time branch -> MBPTA is a probabilistic analysis branch for real-time systems
- must fullfil the requirements of EVT. As discussed in the previous comment from the two requirements only the 1st come from EVT. I propose to change the text as follows: must fullfil certain requirements. In particular ... The first requirement comes from EVT, it can be checked with suitable statistical tests [MBPTA paper] and can be relaxed under some circumstances [Santinelli et al], while the latter... and to the system (hardware/software) properties. Both requirements are necessary ...
- The few previous software are cited in -> The few existing software [Lu],[Lesage] ...
- ... a variant of classical EVT procedures -> a variant of classical MBPTA analysis...
- Dsd in your paper reference-> DSD
- Rtas -> RTAS
- remove &dash from the WCET paper reference
- MBPTA-Cv reference -> MBPTA-CV
