### 2018-06-08 #711 @lkosmid
@federeghe Thank you for the updates in both the code and article.
I will continue with my review and let you know.
