### 2018-06-08 #711 @lkosmid
@federeghe I updated my comments on the short paper.
Some more minor changes are required.
