### 2018-06-11 1d5a842
[Global] Add MACRO to initialize template with 3 arguments
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
