### 2018-06-11 9157757
[AEC/HMI] Cosmetics
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
