### 2018-06-11 6d363c8
[Statistical/Estimator] Add GMLE variant to MLE estimator
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
