+++ src/statistical/estimator_mle.cpp
@@ -40,7 +40,7 @@ namespace chronovise {
 /**
  * @private
  */
-template <typename T_INPUT, typename T_TIME>
+template <typename T_INPUT, typename T_TIME, bool GMLE=false>
 class GEV_Function : public ceres::FirstOrderFunction {
 
 public:
@@ -66,7 +66,7 @@ private:
 
     const MeasuresPool<T_INPUT, T_TIME> &x_data;
 
-    void accumulate_gradient_term(const double* parameters, double *gradient, double z, double u) const;
+    void accumulate_gradient_term(const double* parameters, double *gradient, double z, double u, bool) const;
 
     static bool cmp_0(double x) {
         const double epsilon = 1e-6;
@@ -76,8 +76,40 @@ private:
 
 };
 
-template <typename T_INPUT, typename T_TIME>
-bool GEV_Function<T_INPUT, T_TIME>::Evaluate(const double* parameters, double* cost, double* gradient) const {
+/**
+ * @private
+ */
+template<bool B=false> struct LIKELIHOOD {
+    double operator()(double u, double sg, double xi) {
+        return -std::log(sg) - u + (xi+1.) * std::log(u);
+    }
+};
+
+/**
+ * @private
+ */
+template<> struct LIKELIHOOD<true> {
+    static constexpr double p = 6; // See the paper
+    static constexpr double q = 9; // See the paper
+
+    double operator()(double u, double sg, double xi) {
+        // Generalized maximum-likelihood generalized extreme-value quantile
+        // estimators for hydrologic data, Eduardo S. Martins, Jery R. Stedinger
+        // Water Resources Research, vol. 36, n. 3, pages 737-744, 2000
+
+        const double beta_value = std::tgamma(p) * std::tgamma(q) / std::tgamma(p + q);
+
+        struct LIKELIHOOD<false> l_f;
+        const double mle_f = l_f(u, sg, xi);
+
+        return mle_f
+            + std::pow(0.5 + xi, p-1) * std::pow(0.5 - xi, q-1) / beta_value;
+}
+};
+
+
+template <typename T_INPUT, typename T_TIME, bool GMLE>
+bool GEV_Function<T_INPUT, T_TIME, GMLE>::Evaluate(const double* parameters, double* cost, double* gradient) const {
 
     const double mu = parameters[mu_idx];
     const double sg = parameters[sg_idx];
@@ -119,12 +151,13 @@ bool GEV_Function<T_INPUT, T_TIME>::Evaluate(const double* parameters, double* c
 
         assert(u > 0.);
 
+        struct LIKELIHOOD<GMLE> l_func;
 
         // Ceres will minimize this function, so we have to invert the sign
-        cost_function += - ( -std::log(sg) - u + (xi+1.) * std::log(u) );
+        cost_function += - l_func(u, sg, xi);
 
         if (gradient != NULL) {
-            accumulate_gradient_term(parameters, gradient, z, u);
+            accumulate_gradient_term(parameters, gradient, z, u, GMLE);
         }
     }
 
@@ -139,13 +172,13 @@ bool GEV_Function<T_INPUT, T_TIME>::Evaluate(const double* parameters, double* c
         assert(std::isfinite(gradient[sg_idx]));
         assert(std::isfinite(gradient[xi_idx]));
     }
-    
-    
+
+
     return true;
 }
 
-template <typename T_INPUT, typename T_TIME>
[…]
