### 2018-06-11 e23b370
[Docs] Some paper fixes
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
