+++ docs/paper/paper.bib
@@ -7,7 +7,7 @@
     number    = {03471},
     series    = {Dagstuhl Seminar Proceedings},
     ISSN      = {1862-4405},
-    publisher = {Internationales Begegnungs- und Forschungszentrum f{\"u}r Informatik (IBFI), Schloss Dagstuhl, Germany},
+    publisher = {Internationales Begegnungs- und Forschungszentrum f{\"u}r Informatik ({IBFI}), Schloss Dagstuhl, Germany},
     address   = {Dagstuhl, Germany},
     URL       = {http://drops.dagstuhl.de/opus/volltexte/2004/5}
 }
@@ -22,7 +22,7 @@
 
 @inproceedings{reghenzani2017mixed,
     author={Federico Reghenzani and Giuseppe Massari and William Fornaciari},
-    booktitle={2017 Euromicro Conference on Digital System Design (DSD)},
+    booktitle={2017 Euromicro Conference on Digital System Design ({DSD})},
     title={Mixed Time-Criticality Process Interferences Characterization on a Multicore Linux System},
     year={2017},
     pages={427-434},
@@ -44,7 +44,7 @@
 @inproceedings{bernat2002wcet,
     author={Guillem Bernat and Antoine Colin and Stefan M. Petters},
     booktitle={23rd IEEE Real-Time Systems Symposium, 2002. RTSS 2002.},
-    title={WCET analysis of probabilistic hard real-time systems},
+    title={{WCET} analysis of probabilistic hard real-time systems},
     year={2002},
     volume={},
     number={},
@@ -105,7 +105,7 @@
 
 @inproceedings{santinelli2017revising,
     author={Luca Santinelli and Fabrice Guet and Jerome Morio},
-    booktitle={2017 IEEE Real-Time and Embedded Technology and Applications Symposium (RTAS)},
+    booktitle={2017 IEEE Real-Time and Embedded Technology and Applications Symposium ({RTAS})},
     title={Revising Measurement-Based Probabilistic Timing Analysis},
     year={2017},
     pages={199-208},
@@ -117,7 +117,7 @@
      author = {Reinhard Wilhelm and Jakob Engblom and Andreas Ermedahl and Niklas Holsti and Stephan Thesing
                and David Whalley and Guillem Bernat and Christian Ferdinand and Reinhold Heckmann
                and Tulika Mitra and Frank Mueller and Isabelle Puaut and Peter Puschner and Jan Staschulat and Per Stenstr\"{o}m},
-     title = {The Worst-case Execution-time Problem\&Mdash;Overview of Methods and Survey of Tools},
+     title = {The Worst-case Execution-time Problem. Overview of Methods and Survey of Tools},
      journal = {ACM Transaction of Embededded Computer Systems},
      issue_date = {April 2008},
      volume = {7},
@@ -194,7 +194,7 @@
 
 @misc{abella_jaume_2017_1065776,
   author       = {Jaume Abella},
-  title        = {MBPTA-CV},
+  title        = {{MBPTA-CV}},
   month        = {nov},
   year         = {2017},
   doi          = {10.5281/zenodo.1065776},
@@ -227,4 +227,16 @@
   year={2017},
   school={Universitat Polit{\`e}cnica de Catalunya},
   url={https://www.tdx.cat/handle/10803/460819?locale-attribute=en}
+}
+
+@article{martins2000generalized,
+  title={Generalized maximum-likelihood generalized extreme-value quantile estimators for hydrologic data},
+  author={Martins, Eduardo S and Stedinger, Jery R},
+  journal={Water Resources Research},
+  volume={36},
+  number={3},
+  pages={737--744},
+  year={2000},
+  publisher={Wiley Online Library},
+  doi={10.1029/1999WR900330}
 }
