### 2018-06-11 fcd67de
[Docs] Update estimators in FRAMEWORK.md
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
