### 2018-06-11 23624d3
[Tests] Add one test for GMLE
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
