+++ tests/test_estimator_mle.cpp
@@ -93,7 +93,6 @@ TEST_F(EstimatorMLE_Test, EVT_0_1_M05) {
 
 }
 
-
 TEST_F(EstimatorMLE_Test, EVT_0_1_P05) {
 
     constexpr int n_elem = 200;
@@ -155,6 +154,28 @@ TEST_F(EstimatorMLE_Test, EVT_100_01_025) {
     }
 
     Estimator_MLE<int, double> mle;
+    mle.set_source_evt_approach(&typeid(EVTApproach_BM<int, double>));
+    mle.run(mp);
+    auto res = mle.get_result();
+
+    ASSERT_NEAR(std::dynamic_pointer_cast<GEV_Distribution>(res)->get_location(), 100., 1e-0);    
+    ASSERT_NEAR(std::dynamic_pointer_cast<GEV_Distribution>(res)->get_scale(), 0.1, 1e-0);
+    ASSERT_NEAR(std::dynamic_pointer_cast<GEV_Distribution>(res)->get_shape(), 0.25, 1e-0);
+
+}
+
+TEST_F(EstimatorMLE_Test, GMLE_EVT_100_01_025) {
+
+    constexpr int n_elem = 250;
+    double xs[n_elem] =
+    {100.029585858, 100.052490953, 99.9087113884, 100.097318955, 100.28401529, 99.9280435066, 100.019573272, 99.9766395309, 100.745589297, 100.076558958, 99.9696308853, 99.9355101332, 100.05317911, 100.206675747, 100.2158877, 100.212155041, 99.9561560434, 100.052978117, 100.261623515, 99.9318091534, 100.237418107, 100.301672301, 99.9600164646, 99.9160975329, 100.030389418, 100.213856252, 100.028928081, 100.012703412, 100.039253289, 99.9332398402, 99.9354069597, 100.255433286, 100.074284988, 99.9726931964, 100.247929784, 99.9079804403, 100.025440039, 100.129593102, 99.9922179768, 100.009178757, 100.047110036, 100.291772443, 100.165366099, 99.9129679674, 99.976257699, 100.003136147, 100.247762201, 100.014458203, 99.9659369821, 100.072239707, 100.031973421, 100.299421053, 100.38469746, 100.197370914, 100.122289746, 100.141948187, 100.301550075, 99.9111944467, 99.9913928678, 99.8619026979, 100.206958807, 100.040748443, 99.9995335401, 99.9623870903, 100.049728651, 99.9790836227, 99.9125450264, 99.9192170049, 99.9125316364, 100.011588475, 99.9325790735, 100.021098936, 100.30013623, 99.99614675, 99.9315419667, 100.061638355, 100.261750932, 99.9947843621, 99.8997283223, 99.9385143057, 99.9159427007, 100.140302867, 100.025074968, 100.102005438, 100.117206893, 100.062448393, 100.084623387, 100.2656704, 100.099282811, 100.262291356, 100.028362017, 99.9382078193, 99.9124442138, 100.125173882, 99.9839734919, 100.103369676, 100.094807617, 100.048462358, 100.125646938, 100.039910104, 100.03462371, 100.037700519, 100.38868104, 100.005863532, 99.9305495039, 99.9660828425, 100.109978793, 100.218230224, 100.558656683, 99.959284894, 100.15287662, 100.067120992, 100.009662604, 100.041584079, 100.037010286, 100.094341065, 100.142246782, 99.9823778361, 99.92096203, 100.204904362, 100.005947895, 100.163168668, 99.949369217, 99.9279516193, 100.311943759, 100.263629836, 102.932943333, 100.24722497, 99.8967744238, 100.053068709, 101.40972334, 100.041895473, 100.259612917, 99.913327735, 100.795334497, 100.351098, 100.059902448, 100.017749349, 99.99
[…]
