### 2018-06-11 aa0e7d1
[Docs] Fix grammar paper
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
