### 2018-06-12 federeghe/chronovise#11 @lkosmid
Thank you for the document and code updates, we are almost there :)
Last issues:

On the other hands -> On the other hand
There is are still some capitalisations missing in the references:
Ieee in the RTSS 2012 refernce -> IEEE
Ieee and Sies in SIES 2013 reference-> IEEE and SIES respectively
another Ieee in RTCSA 2010 -> IEEE
Ieee in RTAS 2017 -> IEEE
