### 2018-06-12 fbf5235
[Docs] Fix grammar and capitalization in JOSS paper
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
