+++ docs/paper/paper.bib
@@ -32,7 +32,7 @@
 
 @inproceedings{dasari2013identifying,
     author={Dakshina Dasari and Benny Akesson and Vincent Nélis and Muhammad Ali Awan and Stefan M. Petters},
-    booktitle={2013 8th IEEE International Symposium on Industrial Embedded Systems (SIES)},
+    booktitle={2013 8th {IEEE} International Symposium on Industrial Embedded Systems (SIES)},
     title={Identifying the sources of unpredictability in COTS-based multicore systems},
     year={2013},
     pages={39-48},
@@ -43,7 +43,7 @@
 
 @inproceedings{bernat2002wcet,
     author={Guillem Bernat and Antoine Colin and Stefan M. Petters},
-    booktitle={23rd IEEE Real-Time Systems Symposium, 2002. RTSS 2002.},
+    booktitle={23rd {IEEE} Real-Time Systems Symposium, 2002. RTSS 2002.},
     title={{WCET} analysis of probabilistic hard real-time systems},
     year={2002},
     volume={},
@@ -57,7 +57,7 @@
     author = {Yue Lu and Thomas Nolte and Johan Kraft and Christer Norstrom},
     month = {Aug},
     title = {A Statistical Approach to Response-Time Analysis of Complex Embedded Real-Time Systems},
-    booktitle = {2010 IEEE 16th International Conference on Embedded and Real-Time Computing Systems and Applications},
+    booktitle = {2010 {IEEE} 16th International Conference on Embedded and Real-Time Computing Systems and Applications},
     year = {2010},
     pages = {153-160},
     issn = {2325-1271},
@@ -105,7 +105,7 @@
 
 @inproceedings{santinelli2017revising,
     author={Luca Santinelli and Fabrice Guet and Jerome Morio},
-    booktitle={2017 IEEE Real-Time and Embedded Technology and Applications Symposium ({RTAS})},
+    booktitle={2017 {IEEE} Real-Time and Embedded Technology and Applications Symposium ({RTAS})},
     title={Revising Measurement-Based Probabilistic Timing Analysis},
     year={2017},
     pages={199-208},
