### 2018-06-12 #711 @lkosmid
@federeghe thank you very much for your updates in the code and the manuscript.
We are almost there, some typos only need to be fixed, see my comments.
Regarding the code review, I have some imminent deadlines and back-to-back trips which might reduce my responsiveness and delay a bit the review, but I will try my best to finish it soon.
