### 2018-06-12 federeghe/chronovise#11 @federeghe
Unfortunately it's not simple to notice them without having the possibility to compile markdown+bibtex locally (or at least I didn't find an easy way to do it). Thank you for detecting them.

Fixed :+1:
