### 2018-06-13 federeghe/chronovise#11 @lkosmid
Thank you @federeghe for the quick fix. There is a last Sies remaining and a typo (aimis->aims).
