### 2018-06-14 4b1e712
[Docs] Minor fixes in paper
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
