+++ docs/paper/paper.bib
@@ -32,7 +32,7 @@
 
 @inproceedings{dasari2013identifying,
     author={Dakshina Dasari and Benny Akesson and Vincent Nélis and Muhammad Ali Awan and Stefan M. Petters},
-    booktitle={2013 8th {IEEE} International Symposium on Industrial Embedded Systems (SIES)},
+    booktitle={2013 8th {IEEE} International Symposium on Industrial Embedded Systems ({SIES})},
     title={Identifying the sources of unpredictability in COTS-based multicore systems},
     year={2013},
     pages={39-48},
