### 2018-06-21 #711 @danielskatz
@lkosmid - can we get an update from you on where things are from your point of view?
