### 2018-06-25 #711 @lkosmid
@danielskatz the short paper has been thoroughly reviewed and after several iterations with @federeghe is ready.
From my side it is pending to review the software, its functionality, test environment and documentation (in fact the missing check boxes).
I have already installed the software and gave it a spin a while ago and it seems functional and according to the short paper, but I need to deeply review the code and compare its results with other software packages with similar functionality that I requested e.g. MBPTA-CV
However, as I mentioned I'm between back-to-back business trips and vacations will follow, so I'm afraid that I won't be able to complete those tasks before mid July. I apologise for the long review cycle...
