### 2018-08-05 #711 @danielskatz
Hi @lkosmid - Can I get an update about this review?

I see a bunch of stuff in https://github.com/federeghe/chronovise/issues/11 from a while ago, but it looks like they have all been addressed.  Should that issue now be closed?

Are there other issues that have been raised?
