### 2018-08-06 #711 @lkosmid
Hi @danielskatz , all the issues regarding the content of the short paper paper have been addressed. I had forgotten to close the issue (although I had updated the corresponding checkboxes in the review page) soon after @federeghe updated it.

The only remaining (but time consuming) part from my side is to review the chronovise source code and documentation/examples to verify the claims of the short paper and be able to check the last remaining checkboxes to close the review. I want to make a comparison of some pWCET estimations computed with chronovise with at least MBPTA-CV, which is the reference open source tool in measurement based probabilistic timing analysis. This is necessary in order to make sure that the last claim of the paper, and one of the fundamental contributions of the paper (that the tool can be used by end users like engineers to obtain pWCET estimations with different EVT approaches and compare them) is correct. The problem is that my bandwidth is very limited at the moment, but I have it in high priority. Thank you both for bearing with me and I apologise for the long delay that this causes in  the review process, but I think this extra step will ensure a high-quality review and increase the credibility and impact of the publication.
