### 2018-08-06 #711 @danielskatz
Thanks @lkosmid - I think there are likely some checkboxes in the review that could also be checked at this time (e.g., Authorship), and if you have a chance to do this so it's clear from the list what still needs to be done, that would be great, but is not strictly required - it can wait until you have time to do the code check.
