### 2018-08-10 f9b7de3
[EVT_Approach/CV] Fix vector order in CV estimation
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
