### 2018-08-10 704de90
[Samples] Add a CV values test 2
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
