+++ samples/test_cv_values.h
@@ -1,0 +1,3 @@
+float values[] = {
+31.288456, 30.592342, 32.401566, 32.180346, 29.282594, 29.740145, 31.467475, 30.240663, 32.272663, 28.626454, 30.943366, 30.576685, 32.783710, 27.308996, 30.001495, 30.106990, 25.316190, 32.496190, 35.618464, 29.535630, 30.574064, 29.070791, 30.768636, 29.240345, 29.796444, 33.236302, 28.305540, 28.848234, 29.848174, 29.735057, 32.878619, 34.062526, 28.043597, 28.717953, 31.040951, 29.238600, 33.985704, 33.271740, 31.118038, 31.127275, 29.323207, 31.527295, 32.254885, 30.283142, 33.856268, 31.341276, 29.779778, 30.741925, 32.193189, 29.203674, 29.829075, 34.746575, 29.052158, 31.892682, 31.636323, 33.177942, 31.052024, 25.069605, 28.294944, 31.023471, 30.515550, 33.924895, 32.812630, 30.993667, 30.165661, 26.903044, 33.726319, 30.268053, 26.907995, 30.866565, 30.205908, 28.859308, 30.986124, 28.584974, 27.303950, 26.491397, 29.272382, 28.745812, 30.880301, 26.994718, 29.583529, 26.989749, 33.619482, 29.766055, 32.452418, 30.989886, 31.817190, 27.691562, 32.515111, 32.905065, 25.846377, 29.647024, 28.695836, 30.313068, 28.299548, 30.558092, 31.194187, 30.489433, 29.763307, 27.421184, 30.115127, 28.958134, 27.441946, 29.887546, 29.580639, 30.384203, 31.984665, 31.152488, 32.611568, 28.541388, 28.270851, 29.810158, 32.766428, 32.607363, 29.748376, 28.806252, 26.958707, 31.343683, 30.045689, 27.644423, 29.330844, 26.341090, 30.859979, 29.652013, 33.262485, 27.281144, 31.941013, 30.287280, 30.165207, 31.433328, 32.386968, 27.857876, 32.637804, 27.579994, 27.851770, 28.654880, 31.472924, 27.860079, 30.669430, 30.823765, 30.308240, 31.109141, 27.654309, 32.015173, 30.227041, 31.460102, 28.032976, 30.104063, 31.755199, 32.028282, 29.831303, 26.292937, 27.806359, 30.437255, 31.588492, 30.926248, 28.774740, 34.488800, 30.144695, 31.731029, 29.168607, 27.770124, 31.370505, 32.075346, 33.644423, 28.942023, 26.744067, 33.234605, 30.528273, 27.745708, 28.881648, 28.382310, 32.322009, 31.184211, 30.485495, 30.480955, 28.356999, 31.986224, 30.692789, 29.477733, 29.630574, 29.796530, 28.225914, 31.482754, 32.784416, 34.947801, 31.007838, 28.355039, 30.401964, 27.985895, 28.773657, 28.682073, 28.333541, 30.756359, 27.769315, 31.334483, 31.590666, 32.074984, 29.959138, 31.237906, 33.606127, 30.105987, 29.644220, 33.544965, 24.982393, 29.086822, 34.860755, 29.057092, 28.879343, 27.547065, 31.585901, 25.780119, 28.401210, 31.133950, 29.997114, 31.247849, 30.252882, 31.361736, 27.425168, 30.436170, 26.866866, 31.566564, 29.378860, 31.310649, 28.925065, 30.657287, 32.108144, 26.040601, 26.265224, 26.335289, 31.697184, 30.810375, 28.595088, 32.998092, 30.275590, 26.826357, 27.961709, 27.229533, 31.909773, 
[…]
