+++ samples/CMakeLists.txt
@@ -23,3 +23,11 @@ target_link_libraries(simple_hello_world_cv PUBLIC chronovise)
 if(COVERAGE)
         target_link_libraries(simple_hello_world_cv PUBLIC --coverage)
 endif(COVERAGE)
+
+set(SHW_CV_SOURCES simple_hello_world_cv_2.cpp simple_main_cv.cpp)
+add_executable(simple_hello_world_cv_2 ${SHW_CV_SOURCES})
+target_link_libraries(simple_hello_world_cv_2 PUBLIC chronovise)
+
+#if(COVERAGE)
+#        target_link_libraries(simple_hello_world_cv_2 PUBLIC --coverage)
+#endif(COVERAGE)
