### 2018-08-10 #711 @federeghe
Dear @lkosmid,

To simplify your review and reduce your effort, I pushed a small example over 10000 samples generated from a normal distribution for MBPTA-CV together with a small bugfix. The example is called `simple_hello_world_cv_2.cpp` and in the samples folder you can find the original dataset (both in txt for R and in .h for chronovise).
I hope that this can be useful.

The result of the original MBPTA-CV R software is:
```
CV-Value = 0.981919
pWCET(10-6) = 42
pWCET(10-9) = 46
```
The result of chronovise is:
```
CV-Value = 0.986187
pWCET(10-6) = 42.8073
pWCET(10-9) = 46.9047
```
