### 2018-08-10 ad49547
[Build] Fix Coverage linking
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
