### 2018-08-20 #711 @lkosmid
Hi @danielskatz and @federeghe , I was out of the office last week.
@federeghe thank you for the example you added, it definitely simplifies a lot  my review.
I will be able to provide an update this week.
