### 2018-08-24 federeghe/chronovise#15 @lkosmid
Utility function to read samples from a file
In the [simple_hello_world_cv_2.cpp](https://github.com/federeghe/chronovise/blob/master/samples/simple_hello_world_cv_2.cpp) example the time samples are provided statically in an array defined in the [test_cv_values.h](https://github.com/federeghe/chronovise/blob/master/samples/test_cv_values.h) header file. However, this means that the program has to be recompiled every time in order to be used with a new data set.

In order chronovise to be easier to use by end-users (e.g. engineers not developers) , it would be useful to provide a utility function that can parse directly an input file such as [test_cv_values.txt](https://github.com/federeghe/chronovise/blob/master/samples/test_cv_values.txt). Moreover, this will provide compatibility with MBPTA-CV.
