### 2018-08-24 federeghe/chronovise#15 @federeghe
You're right, this was a very ugly hack just for the review. However, in C++ it's very easy to read numbers from a file using `ifstream`, declaring somewhere (let's say global)

```
std::ifstream my_input_file ("file.txt");
```

and then in `onRun()` doing something like this:

```
float num;
my_input_file >> num;
this->add_sample(num);
```

I didn't put too much effort in that example, because the user is required to implement the class `SimpleExecutionContext` or `ExecutionContext` as he or she desires, but I can fix this horrible hack.
