### 2018-08-24 1694bed
[Samples] Improve example for MBPTA-CV
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
