### 2018-08-24 federeghe/chronovise#15 @lkosmid
What I was thinking is that instead of modifying the example, to overload add_sample() in order to get a filename as an argument, since this might be a frequently used feature.
