### 2018-08-24 federeghe/chronovise#16 @lkosmid
JOSS Submission Documentation Enchancements
In general the provided documentation is in a very good shape. I think that the page about the brief information about MBPTA is very useful for any end user that is going to use it.

I only have some small suggestions for improvement:

- The member functions in the SimpleExecutionContext class miss doxyzen documentation
-  The Sample applications documentation page should be expanded to cover the new example [simple_hello_world_cv_2.cpp](https://github.com/federeghe/chronovise/blob/master/samples/simple_hello_world_cv_2.cpp). It doesn't need to be long, just to mention that it shows how you can feed measurement samples to chronovise, preferably read from a file once the feature #15 is implemented.
