### 2018-08-24 #711 @lkosmid
I think that the documentation is in a very good state. I only have a minor suggestion for improving the documentation ([#16](https://github.com/federeghe/chronovise/issues/16)) and one small feature to be implemented ([#15](https://github.com/federeghe/chronovise/issues/15)).
