### 2018-08-27 4a6a573
[Utility] Add FileImporter class
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
