### 2018-08-27 d456e50
[AEC] Small fixes and docs
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
