### 2018-08-27 a4cc916
[SEC+Samples] Add add_sample_from_file() method and example of usage
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
