+++ samples/simple_hello_world_cv_2.cpp
@@ -14,9 +14,6 @@ using namespace chronovise;
 
 using exit_code_t = AbstractExecutionContext<unsigned int, double>::exit_code_t;
 
-#include "test_cv_values.h"
-
-static std::ifstream file_test_values ("test_cv_values.txt");
 
 SimpleHelloWorld::SimpleHelloWorld() noexcept {
     // Nothing to do...
@@ -53,18 +50,13 @@ exit_code_t SimpleHelloWorld::onSetup() noexcept {
     // Select the estimator. Currently the only one available is the Maximum-Likelihood-Estimator
     this->use_estimator_CV();
 
-    // Select the Kolmogorov-Smirnov test as test on the estimated probability distribution for
-    // pWCET
-//    this->use_KS_as_post_evt_test(0.01);
 
     // Let's print some debug information
     this->print_configuration_info();
     this->print_legend();
 
-    if (!file_test_values) {
-        std::cout << "** Unable to open the input file" << std::endl;
-        return AEC_GENERIC_ERROR;
-    }
+    // Gather input values from file
+    this->use_file_as_source("test_cv_values.txt");
 
     return AEC_OK;
 }
@@ -81,17 +73,27 @@ exit_code_t SimpleHelloWorld::onConfigure() noexcept
 
 exit_code_t SimpleHelloWorld::onRun() noexcept {
 
-    double value;
-    file_test_values >> value;
+//    std::cout << get_iteration() << std::endl;
 
-    this->add_sample(value);
+    try {
+
+        // Get a sample from the file specified
+        // in onSetup().
+        this->add_sample_from_file();
+
+    } catch( std::ios::failure ex) {
+        std::cout << std::endl << "INPUT ERROR: "
+                  << ex.what() << std::endl;
+        return AEC_INPUT_ERROR;
+    }
+
 
     return AEC_OK;
 }
 
 exit_code_t SimpleHelloWorld::onMonitor() noexcept {
 
-    if (get_iteration() > 10000) {
+    if (get_iteration() >= 10000) {
         return AEC_OK;
     } else {
         return AEC_CONTINUE;
