### 2018-08-27 17c8876
[SEC] Add documentation
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
