+++ include/sec.hpp
@@ -47,12 +47,26 @@ class SimpleExecutionContext : public AbstractExecutionContext<T_INPUT, T_TIME> 
 
 public:
 
+    /**
+     * The construction of SEC class does not perform anything special, check the
+     * parent class constructor.
+     * @see AbstractExecutionContext::AbstractExecutionContext()
+     */
     explicit SimpleExecutionContext() : AbstractExecutionContext<T_INPUT, T_TIME>()
     {
     }
 
+    /**
+     * A default destructor.
+     */
     virtual ~SimpleExecutionContext() = default;
 
+    /**
+     * This method add the LjungBox test as sample test with the given parameter.
+     * @param significance_level The level of significance (\alpha) of the test
+     * @param n_lags The number of lags to use
+     * @see TestLjungBox::TestLjungBox()
+     */
     void use_LjungBox_as_sample_test(double significance_level, int n_lags) {
         std::shared_ptr<StatisticalTest<T_INPUT, T_TIME>> stat_test(
             new TestLjungBox<T_INPUT, T_TIME>(significance_level, n_lags)
@@ -60,6 +74,14 @@ public:
         this->add_sample_test(stat_test);
     }
 
+    /**
+     * Initialize the EVT approach as Block-Maxima with the specified
+     * block size and the provided ratio trainset/testset.
+     * @param block_size The block size to use
+     * @param ratio_train_test The ratio of samples
+     * @see EVTApproach_BM::EVTApproach_BM()
+     * @see AbstractExecutionContext::set_evt_approach()
+     */
     void use_evt_approach_BM(int block_size, float ratio_train_test) {
         std::unique_ptr<EVTApproach<T_INPUT, T_TIME>> evt_app(
             new EVTApproach_BM<T_INPUT, T_TIME>(block_size)
@@ -68,6 +90,14 @@ public:
         is_BM = true;
     }
 
+    /**
+     * Initialize the EVT approach as Peak-over-Threshold with the specified
+     * block size and the provided ratio trainset/testset.
+     * @param threshold The threshold to use
+     * @param ratio_train_test The ratio of samples
+     * @see EVTApproach_PoT::EVTApproach_PoT()
+     * @see AbstractExecutionContext::set_evt_approach()
+     */
     void use_evt_approach_PoT(T_TIME threshold, float ratio_train_test) {
         std::unique_ptr<EVTApproach<T_INPUT, T_TIME>> evt_app(
             new EVTApproach_PoT<T_INPUT, T_TIME>(threshold)
@@ -76,6 +106,12 @@ public:
         is_BM = false;
     }
 
+    /**
+     * Initialize the EVT approach as MBPTA-CV method.
+     * @param ratio_train_test The ratio of samples
+     * @see EVTApproach_CV::EVTApproach_CV()
+     * @see AbstractExecutionContext::set_evt_approach()
+     */
     void use_evt_approach_CV(float ratio_train_test) {
         std::unique_ptr<EVTApproach<T_INPUT, T_TIME>> evt_app(
             new EVTApproach_CV<T_INPUT, T_TIME>()
@@ -84,6 +120,10 @@ public:
         is_BM = false;
     }
 
+    /**
+     * Set the estimator to be the Maximum-Likelihood-Estimator
+     * @see Estimator_MLE::Estimator_MLE()
+     */
     void use_estimator_MLE() {
         std::unique_ptr<Estimator<T_INPUT, T_TIME>> evt_est(
             new Estimator_MLE<T_INPUT, T_TIME>()
@@ -91,6 +131,10 @@ public:
         this->set_evt_estimator(std::move(evt_est));
     }
 
+    /**
+     * Set the estimator to be the Generalized Maximum-Likelihood-Estimator
[…]
