### 2018-08-27 b4b6ba7
[Samples] Remove old unused file
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
