### 2018-08-27 e0c962a
[Docs] Add CV version of samples in documentation
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
