### 2018-08-27 federeghe/chronovise#15 @federeghe
Ok, I created an utility to read files (4a6a5730)  and a function add_sample_from_file() in SimpleExecutionContext (a4cc916ec). The example has been modified accordingly. I didn't use the overload to avoid name hiding problem of the parent class methods.
