### 2018-08-27 federeghe/chronovise#16 @federeghe
Both done in 17c8876 and e0c962a, thank you!
