### 2018-08-27 #711 @federeghe
@lkosmid Thank you for the suggestions and the detailed comments, issues federeghe/chronovise#15 and federeghe/chronovise#16 have been fixed.
