### 2018-08-27 #711 @lkosmid
@federeghe Thank you for applying the suggestions! I only have a very minor comment for [#16](https://github.com/federeghe/chronovise/issues/16)
