### 2018-08-27 9fcd526
[Docs] Improve documentation for SAMPLES
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
