### 2018-08-27 0aac975
[Docs] Minor fix
Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>
