### 2018-08-27 #711 @lkosmid
@federeghe Thank you for addressing all the comments. I don't have anything else, the review is complete. I have checked the last remaining item in the above checklist.
@danielskatz What are the next steps?
