### 2018-08-27 #711 @danielskatz
Thanks very much @lkosmid

@federeghe , please archive the current version of the repository in zenodo or similar, and let me know the DOI.
